api/requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
pymongo==4.11.3
python-dotenv==1.0.1
pydantic==2.10.6
httpx==0.28.1 